+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH6DHF8PFJG84C1PC2YA2QTY"
+        "01KH6K3N7307MZCF2E3EQWXNBG"
     )
     r"""The unique identifier of the data source"""