+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH6DHF8QCR0GSB6C5JZ8F3VN"
+        "01KH6K3N760Z08DTMYR1ZTVMJY"
     )
     r"""The unique identifier of the data source"""