+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH6DHF8VR9V0X7RCSK8EWZJC"
+        "01KH6K3N7ASCSTRT2EQRXRY5SK"
     )
     r"""The unique identifier of the data source"""