+++ packages/orq-rc/RELEASES.md
@@ -5539,3 +5539,13 @@ Based on:
 - [python v4.3.0-rc.54] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.54] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.54 - packages/orq-rc
+
+## 2026-02-11 14:56:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.710.0 (2.818.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.55] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.55] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.55 - packages/orq-rc